+++ src/main/java/com/heartsave/todaktodak_api/common/exception/ErrorResponse.java
@@ -1,6 +1,5 @@
 package com.heartsave.todaktodak_api.common.exception;
 
-import com.heartsave.todaktodak_api.common.type.ErrorSpec;
 import java.util.Map;
 import lombok.*;
